NFC: Split out `TypeCheckAccessNotes.cpp`
